Rework the structure to support conn unicast function returning interface type
